Disable mapping of displayNames
Mapping of German displayNames leads to loads of errors because
the validator validates against English displayNames. Therefore, the
mapping of displayNames has been temporarily disabled.

For an Austrian Lab IG it is assumed that German displayNames will be
used in value sets. And for cross border use cases the displayNames
will have to be translated by the NCPeH.